+++ camera-ui-coral/src/main.py
@@ -196,7 +196,7 @@ class CoralPlugin(BasePlugin, ObjectDetectionInterface):
     async def _add_sensors(self, camera: CameraDevice) -> None:
         sensors: dict[str, Any] = {}
 
-        obj = CoralObjectSensor(self, self.logger)
+        obj = CoralObjectSensor(self, camera, self.logger)
         await camera.addSensor(obj)
         sensors["object"] = obj